LayoutTests/crypto/subtle/decrypt-malformed-parameters-expected.txt

CONSOLE MESSAGE: AES-CBC and AES-CTR do not provide authentication by default, and implementing it manually can result in minor, but serious mistakes. We recommended using authenticated encryption like AES-GCM to protect against chosen-ciphertext attacks.
Test decrypting with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.decrypt() rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.decrypt(1) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.decrypt(1, 2) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.decrypt(aesCbcParams, wrongKey, cipherText) rejected promise  with InvalidAccessError: CryptoKey doesn't match AlgorithmIdentifier.
PASS crypto.subtle.decrypt(aesCbcParams, wrongKey, cipherText) rejected promise  with InvalidAccessError: CryptoKey doesn't support decryption.
PASS successfullyParsed is true

TEST COMPLETE
